Remove task chain finalization duplicate code
